Check that winnfsd exists before installing service

+++ scripts/windows_ddev_nfs_setup.sh
@@ -26,6 +26,10 @@ if [ "$status" = "SERVICE_STOPPED" ] ; then
     exit 0
 fi
 
+if [ ! -f "/C/Program Files/ddev/winnfsd.exe" ] ; then
+    echo "winnfsd is not installed in C:\Program Files\ddev\ - Plesae use the ddev windows isntaller to get it installed, or do a manual installation."
+    exit 101
+fi
 sudo nssm install nfsd "C:\Program Files\ddev\winnfsd.exe" -id ${DDEV_WINDOWS_UID} ${DDEV_WINDOWS_GID} -log off "C:\\Users"
 sudo nssm start nfsd
